requirements.txt
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cryptography==3.1
decorator==4.4.2
deepspeech==0.8.2
idna==2.10
numpy==1.19.2
pycparser==2.20
PyQt5==5.15.1
PyQt5-sip==12.8.1
python-telegram-bot==12.8
requests==2.24.0
six==1.15.0
tornado==6.0.4
urllib3==1.25.10
webrtcvad==2.0.10
